pyproject.toml
[project]
name = "retro68-serial-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyserial>=3.5",
]
